str: switch to using _gnutls_buffer_pop_uint* where possible
As _gnutls_buffer_pop_prefix*(..., 0) is equivalent to
_gnutls_buffer_pop_uint*(...), use the latter instead.

After that, only callers of _gnutls_buffer_pop_prefix*(..., 1) are in
_gnutls_buffer_pop_datum_prefix* in str.c. Replace them with
_gnutls_buffer_pop_uint*(...) and manual check after that.

+++ lib/gnutls_int.h
@@ -1126,7 +1126,7 @@ typedef struct {
 	uint32_t lifetime;
 	uint32_t age_add;
 	uint8_t nonce[255];
-	size_t nonce_size;
+	uint8_t nonce_size;
 	const mac_entry_st *prf;
 	uint8_t resumption_master_secret[MAX_HASH_SIZE];
 	gnutls_datum_t ticket;
